Fix terminal-index wrap and snapshot-offset indexing in analysis curves
Address review findings on PR #1618:
- CumulativePnL: use long arithmetic for mark-to-market loop start so an
  open position entered on Integer.MAX_VALUE does not wrap to MIN_VALUE
- InvestedInterval: iterate with a long index so intervals ending at the
  terminal bar stop instead of wrapping past the window
- OmegaRatioCriterion: derive begin/end indexes and return-rate positions
  from the captured snapshot instead of live series accessors, and scan
  with a long index to avoid wrapping at Integer.MAX_VALUE
- ConcurrentBarSeries: add @since tag to snapshot() override Javadoc
- BaseBarSeriesTest: assert snapshot bar data is unmodifiable
- New regressions: terminal open position, terminal interval, terminal
  Omega scan, snapshot-offset indexing; formatting aligned with spotless

+++ ta4j-core/src/main/java/org/ta4j/core/analysis/CumulativePnL.java
@@ -179,12 +179,12 @@ public final class CumulativePnL implements PerformanceIndicator {
 
         if (equityCurveMode == EquityCurveMode.MARK_TO_MARKET) {
             Num averageCostPerPeriod = averageHoldingCostPerPeriod(position, endIndex, numFactory);
-            int start = Math.max(entryIndex + 1, seriesBegin + 1);
-            for (int i = start; i < endIndex; i++) {
-                Num close = barSeries.getBar(i).getClosePrice();
+            long start = Math.max((long) entryIndex + 1, (long) seriesBegin + 1);
+            for (long i = start; i < endIndex; i++) {
+                Num close = barSeries.getBar((int) i).getClosePrice();
                 Num netIntermediate = addCost(close, averageCostPerPeriod, isLong);
                 Num delta = isLong ? netIntermediate.minus(netEntryPrice) : netEntryPrice.minus(netIntermediate);
-                addValue(i, delta);
+                addValue((int) i, delta);
             }
             Num exitRaw = resolveExitPrice(position, endIndex, barSeries);
             Num netExit = addCost(exitRaw, averageCostPerPeriod, isLong);
